requirements.txt
aniso8601==9.0.1
appnope==0.1.2
argon2-cffi==21.1.0
astroid==2.8.5
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cffi==1.15.0
click==8.0.3
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
filelock==3.3.0
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-mongoengine==1.0.0
flask-restx==0.5.1
Flask-WTF==1.0.0
idna==3.3
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.23.1
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pymongo==3.12.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
Send2Trash==1.8.0
six==1.16.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
typing_extensions==4.0.0
virtualenv==20.8.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.13.3
WTForms==3.0.0
